+++ rest/src/main/java/com/ericsson/bss/cassandra/ecchronos/rest/OnDemandRepairManagementRESTImpl.java
@@ -140,11 +140,13 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
             final boolean all,
             @RequestParam(required = false, defaultValue = "false")
             @Parameter(description = "Force repair of TWCS tables, which are normally ignored.")
-            final boolean forceRepairTWCS)
-
+            final boolean forceRepairTWCS,
+            @RequestParam(required = false, defaultValue = "false")
+            @Parameter(description = "Force repair of tables disabled in the schedule .")
+            final boolean forceRepairDisabled)
     {
         return ResponseEntity.ok(runOnDemandRepair(nodeID, keyspace, table, getRepairTypeOrDefault(repairType), all,
-                forceRepairTWCS));
+                forceRepairTWCS, forceRepairDisabled));
     }
 
     private RepairType getRepairTypeOrDefault(final RepairType repairType)
@@ -207,7 +209,7 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
             final String keyspace, final String table,
             final RepairType repairType,
             final boolean all,
-            final boolean forceRepairTWCS)
+            final boolean forceRepairTWCS, final boolean forceRepairDisabled)
     {
         try
         {
@@ -219,11 +221,13 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
 
             if (keyspace != null)
             {
-                onDemandRepairs = getOnDemandRepairsForKeyspace(keyspace, table, repairType, nodeUUID, forceRepairTWCS);
+                onDemandRepairs = getOnDemandRepairsForKeyspace(keyspace, table, repairType,
+                        nodeUUID, forceRepairTWCS, forceRepairDisabled);
             }
             else
             {
-                onDemandRepairs = runLocalOrCluster(nodeUUID, repairType, myTableReferenceFactory.forCluster(), forceRepairTWCS);
+                onDemandRepairs = runLocalOrCluster(nodeUUID, repairType, myTableReferenceFactory.forCluster(),
+                        forceRepairTWCS, forceRepairDisabled);
             }
             return onDemandRepairs;
         }
@@ -249,7 +253,8 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
                                                                final String table,
                                                                final RepairType repairType,
                                                                final UUID nodeUUID,
-                                                               final boolean forceRepairTWCS) throws EcChronosException
+                                                               final boolean forceRepairTWCS,
+                                                               final boolean forceRepairDisabled) throws EcChronosException
     {
         List<OnDemandRepair> onDemandRepairs;
         if (table != null)
@@ -265,13 +270,19 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
                 throw new ResponseStatusException(BAD_REQUEST,
                         "Table " + keyspace + "." + table + " uses TWCS");
             }
+            if (!myOnDemandRepairScheduler.checkTableEnabled(tableReference, forceRepairDisabled))
+            {
+                throw new ResponseStatusException(BAD_REQUEST,
+                        "Table " + keyspace + "." + table + " is disabled");
+            }
             onDemandRepairs = runLocalOrCluster(nodeUUID, repairType,
-                    Collections.singleton(myTableReferenceFactory.forTable(keyspace, table)), forceRepairTWCS);
+                    Collections.singleton(myTableReferenceFactory.forTable(keyspace, table)),
+                    forceRepairTWCS, forceRepairDisabled);
         }
         else
         {
             onDemandRepairs = runLocalOrCluster(nodeUUID, repairType,
-                    myTableReferenceFactory.forKeyspace(keyspace), forceRepairTWCS);
+                    myTableReferenceFactory.forKeyspace(keyspace), forceRepairTWCS, forceRepairDisabled);
         }
         return onDemandRepairs;
     }
@@ -301,18 +312,19 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
             final UUID nodeID,
             final RepairType repairType,
             final Set<TableReference> tables,
-            final boolean forceRepairTWCS)
+            final boolean forceRepairTWCS,
+            final boolean forceRepairDisabled)
             throws EcChronosException
     {
         if (nodeID == null)
         {
-            return runForCluster(repairType, tables, forceRepairTWCS);
+            return runForCluster(repairType, tables, forceRepairTWCS, forceRepairDisabled);
         }
         List<OnDemandRepair> onDemandRepairs = new ArrayList<>();
         Node node = myDistributedNativeConnectionProvider.getNodes().get(nodeID);
         for (TableReference tableReference : tables)
         {
-            if (!rejectForTWCS(tableReference, forceRepairTWCS) && myReplicatedTableProvider.accept(node, tableReference.getKeyspace()))
+            if (isRepairableTable(forceRepairTWCS, forceRepairDisabled, tableReference, node))
             {
                 onDemandRepairs.add(new OnDemandRepair(
                         myOnDemandRepairScheduler.scheduleJob(tableReference, repairType, nodeID)));
@@ -320,10 +332,17 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
         }
         return onDemandRepairs;
     }
+    private boolean isRepairableTable(final boolean forceRepairTWCS, final boolean forceRepairDisabled, final TableReference tableReference, final Node node)
+    {
+        return !rejectForTWCS(tableReference, forceRepairTWCS)
+                && myReplicatedTableProvider.accept(node, tableReference.getKeyspace())
+                && myOnDemandRepairScheduler.checkTableEnabled(tableReference, forceRepairDisabled);
+    }
     private List<OnDemandRepair> runForCluster(
             final RepairType repairType,
             final Set<TableReference> tables,
-            final boolean forceRepairTWCS)
+            final boolean forceRepairTWCS,
+            final boolean forceRepairDisabled)
             throws EcChronosException
     {
         List<OnDemandRepair> onDemandRepairs = new ArrayList<>();
@@ -332,7 +351,7 @@ public class OnDemandRepairManagementRESTImpl implements OnDemandRepairManagemen
             Collection<Node> availableNodes = myDistributedNativeConnectionProvider.getNodes().values();
             for (Node eachNode : availableNodes)
             {
-                if (!rejectForTWCS(tableReference, forceRepairTWCS) && myReplicatedTableProvider.accept(eachNode, tableReference.getKeyspace()))
+                if (isRepairableTable(forceRepairTWCS, forceRepairDisabled, tableReference, eachNode))
                 {
                     onDemandRepairs.add(new OnDemandRepair(
                             myOnDemandRepairScheduler.scheduleJob(tableReference, repairType, eachNode.getHostId())));